requirements.txt
torch
pandas
transformers
langchain
pyttsx3
speechrecognition
textblob
protobuf==3.20.3
google-generativeai
PyAudio
deep_translator
python-dotenv
gtts
pygame
google-api-python-client
google-auth
python-dateutil
pytz
